bot/requirements.txt
alpaca-py
requests
pandas
numpy